+++ src/uni_modules/uview-plus/types/comps/tooltip.d.ts
@@ -25,10 +25,14 @@ declare interface TooltipProps {
    */
   bgColor?: string
   /**
+   * 弹出提示框的背景色
+   */
+  popupBgColor?: string
+  /**
    * 弹出提示的方向
    * @default "top"
    */
-  direction?: 'top' | 'bottom'
+  direction?: 'top' | 'bottom' | 'left' | 'right'
   /**
    * 弹出提示的z-index，nvue无效
    * @default 10071
@@ -54,10 +58,41 @@ declare interface TooltipProps {
    */
   showToast?: boolean
   /**
+   * 触发方式
+   * @default "longpress"
+   */
+  triggerMode?: 'longpress' | 'click' | 'manual'
+  /**
+   * 强制定位
+   */
+  forcePosition?: Record<string, string | number>
+  /**
+   * 是否显示，triggerMode为manual时使用
+   * @default false
+   */
+  show?: boolean
+  /**
+   * 是否开启单例模式，开启该属性的tooltip同一页面同时只显示一个
+   * @default false
+   */
+  singleton?: boolean
+  /**
    * 点击触发事件
    * @param index 被点击按钮的索引
    */
   onClick?: (index: number) => any
+  /**
+   * 显示状态变化时触发
+   */
+  ['onUpdate:show']?: (value: boolean) => any
+  /**
+   * 打开时触发
+   */
+  onOpen?: () => any
+  /**
+   * 关闭时触发
+   */
+  onClose?: () => any
 }
 
 declare interface _Tooltip {